test(mcp): the split-surface population is the repo's projects, not the tool surface's closure
Round 1: the arm that catches a tool declared outside the assembly the server
registers from walked OUTWARD from that assembly through its references. A
closure reaches only what its root references, and the reference direction is
housecarl-generator -> housecarl-mcp -> housecarl-core, so housecarl-generator
was never in the population. It is also the one shipped assembly that already
compiles against the MCP SDK, so a marked type there needs no csproj change:
the reviewer added a real [McpServerToolType] tool to it, and the census tests
and ci-all were both green over an unreachable tool. That is exactly the shape
the arm exists to catch, and the comment above it said the opposite -- "latent
today, because housecarl-core carries no MCP SDK reference".

The population is now every project under src/, from the csproj files, less this
test project, which is not shipped and declares the offender fixture on purpose.
Same derivation the roster's second spelling already used, so the two walks share
one home in RepoProjects instead of spelling the assembly load twice.

That move also fixes the load: it enumerated <project>/bin directly, which throws
DirectoryNotFoundException on a project that has never been built -- the one case
the written refusal was for, so the sentence was unreachable and the reader got a
bare path exception. It now enumerates the project directory, which exists by
construction, and filters to the bin tree.

Refs #478
